+++ local/plan/objectivescales/view.php
@@ -19,6 +19,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author Alastair Munro <alastair@catalyst.net.nz>
+ * @author Simon Coggins <simonc@catalyst.net.nz>
  * @package totara
  * @subpackage plan
  */
@@ -26,9 +27,6 @@
 require_once('../../../config.php');
 require_once($CFG->libdir.'/adminlib.php');
 require_once('lib.php');
-require_js(array(
-    $CFG->wwwroot.'/local/js/lib/jquery-1.3.2.min.js',
-));
 
 ///
 /// Setup / loading data
@@ -52,9 +50,19 @@ if (!$objective = get_record('dp_objective_scale', 'id', $id)) {
     error(get_string('error:objectivescaleidincorrect', 'local_plan'));
 }
 
+$scale_used = dp_objective_scale_is_used($id);
+
 if($delete) {
     if(!$value = get_record('dp_objective_scale_value', 'id', $delete)) {
        print_error('error:invalidobjectivescalevalueid', 'local_plan');
+    }
+
+    if($scale_used) {
+        print_error('error:nodeleteobjectivescalevalueinuse', 'local_plan');
+    }
+
+    if($value->id == $objective->defaultid) {
+        print_error('error:nodeleteobjectivescalevaluedefault', 'local_plan');
     }
 
     if($confirm) {
@@ -75,7 +83,7 @@ if($delete) {
         }
 
         commit_sql();
-        totara_set_notification(get_string('deletedobjectivescalevalue', 'local_plan'), $CFG->wwwroot.'/local/plan/objectivescales/view.php?id='.$objective->id, array('style' => 'notifysuccess'));
+        totara_set_notification(get_string('deletedobjectivescalevalue', 'local_plan', format_string($objective->name)), $CFG->wwwroot.'/local/plan/objectivescales/view.php?id='.$objective->id, array('style' => 'notifysuccess'));
 
     } else {
         $returnurl = "{$CFG->wwwroot}/local/plan/objectivescales/view.php?id={$objective->id}";
@@ -112,7 +120,7 @@ $str_set = get_string('set', 'local_plan');
 if ((!empty($moveup) or !empty($movedown))) {
 
     // Can't reorder a scale that's in use
-    if ( dp_objective_scale_is_used($objective->id) ) {
+    if ( $scale_used ) {
         $returnurl = "{$CFG->wwwroot}/local/plan/objectivescales/view.php?id={$objective->id}";
         print_error('error:noreorderobjectiveinuse', 'local_plan', $returnurl);
     }
@@ -180,6 +188,7 @@ if ($default) {
     if (!update_record('dp_objective_scale', $s)) {
         error(get_string('error:updateobjectivescale','local_plan'));
     } else {
+        totara_set_notification(get_string('objectivescaledefaultupdated', 'local_plan'), null, array('style' => 'notifysuccess'));
         // Fetch the update scale record so it'll show up to the user.
         $objective = get_record('dp_objective_scale', 'id', $id);
     }
@@ -201,9 +210,23 @@ $navlinks[] = array('name'=>format_string($objective->name), 'link'=>'', 'type'=
 
 admin_externalpage_print_header('', $navlinks);
 
+print_single_button($CFG->wwwroot . '/local/plan/objectivescales/index.php', null, get_string('allobjectivescales', 'local_plan'));
+
 // Display info about scale
-print_heading(format_string($objective->name), 'left', 1);
+print_heading(get_string('objectivescalex', 'local_plan', format_string($objective->name)), 'left', 1);
 echo '<p>'.format_string($objective->description, FORMAT_HTML).'</p>';
+
+// Display warning if scale is in use
+if($scale_used) {
+    print_container(get_string('objectivescaleinuse', 'local_plan'), true, 'notifysuccess');
+}
+
+// Display warning if proficient values don't make sense
+$max_achieved = get_field('dp_objective_scale_value', 'MAX(sortorder)', 'achieved', 1, 'objscaleid', $id);
+$min_notachieved = get_field('dp_objective_scale_value', 'MIN(sortorder)', 'achieved', 0, 'objscaleid', $id);
+if(isset($max_achieved) && isset($min_notachieved) && $max_achieved > $min_notachieved) {
+    print_container(get_string('nonsensicalachievedvalues', 'local_plan'), true, 'notifyproblem');
+}
 
 // Display objective scale values
 if ($values) {
@@ -222,7 +245,8 @@ if ($values) {
         helpbutton('objectivescaledefault', 'Help with Default Value', 'local_plan', true, false, '', true);
     $table->align[] = 'center';
 
-    $table->head[] = get_string('achieved', 'local_plan');
+    $table->head[] = get_string('achieved', 'local_plan').' '.
+        helpbutton('objectivescalevalueachieved', 'Help with Proficient Value', 'local_plan', true, false, '', true);
     $table->align[] = 'center';
 
     $table->head[] = get_string('edit');
@@ -242,10 +266,12 @@ if ($values) {
         $buttons = array();
 
         // Is this the default value?
+        $disabled = ($numvalues == 1) ? ' disabled="disabled"' : '';
         if ($value->id == $objective->defaultid) {
-            $row[] = '<input type="radio" name="default" value="'.$value->id.'" checked="checked" />';
+            $row[] = '<input type="radio" name="default" value="'.$value->id.'" checked="checked" ' .
+                $disabled . ' />';
         } else {
-            $row[] = '<input type="radio" name="default" value="'.$value->id.'" />';
+            $row[] = '<input type="radio" name="default" value="'.$value->id.'" ' . $disabled . ' />';
         }
 
         if ($value->achieved) {
@@ -257,11 +283,19 @@ if ($values) {
         $buttons[] = "<a href=\"{$CFG->wwwroot}/local/plan/objectivescales/editvalue.php?id={$value->id}\" title=\"$str_edit\">".
             "<img src=\"{$CFG->pixpath}/t/edit.gif\" class=\"iconsmall\" alt=\"$str_edit\" /></a>";
 
-        $buttons[] = "<a href=\"{$CFG->wwwroot}/local/plan/objectivescales/view.php?id={$objective->id}&amp;delete={$value->id}\" title=\"$str_delete\">".
-            "<img src=\"{$CFG->pixpath}/t/delete.gif\" class=\"iconsmall\" alt=\"$str_delete\" /></a>";
+        if(!$scale_used) {
+            if($value->id == $objective->defaultid) {
+                $buttons[] = "<img src=\"{$CFG->pixpath}/t/dismiss.gif\" class=\"iconsmall\" alt=\"" . get_string('error:nodeleteobjectivescalevaluedefault', 'local_plan') . "\" title=\"" . get_string('error:nodeleteobjectivescalevaluedefault', 'local_plan') . "\" /></a>";
+            } else {
+                $buttons[] = "<a href=\"{$CFG->wwwroot}/local/plan/objectivescales/view.php?id={$objective->id}&amp;delete={$value->id}\" title=\"$str_delete\">".
+                    "<img src=\"{$CFG->pixpath}/t/delete.gif\" class=\"iconsmall\" alt=\"$str_delete\" /></a>";
+            }
+        } else {
+            $buttons[] = "<img src=\"{$CFG->pixpath}/t/dismiss.gif\" class=\"iconsmall\" alt=\"" . get_string('error:nodeleteobjectivescalevalueinuse', 'local_plan') . "\" title=\"" . get_string('error:nodeleteobjectivescalevalueinuse', 'local_plan') . "\" /></a>";
+        }
 
         // If value can be moved up
-        if ($count > 1) {
+        if ($count > 1 && !$scale_used) {
             $buttons[] = "<a href=\"{$CFG->wwwroot}/local/plan/objectivescales/view.php?id={$objective->id}&moveup={$value->id}\" title=\"$str_moveup\">".
                 "<img src=\"{$CFG->pixpath}/t/up.gif\" class=\"iconsmall\" alt=\"$str_moveup\" /></a>";
         } else {
@@ -269,7 +303,7 @@ if ($values) {
         }
 
         // If value can be moved down
-        if ($count < $numvalues) {
+        if ($count < $numvalues && !$scale_used) {
             $buttons[] = "<a href=\"{$CFG->wwwroot}/local/plan/objectivescales/view.php?id={$objective->id}&movedown={$value->id}\" title=\"$str_movedown\">".
                 "<img src=\"{$CFG->pixpath}/t/down.gif\" class=\"iconsmall\" alt=\"$str_movedown\" /></a>";
         } else {
@@ -281,31 +315,18 @@ if ($values) {
         $table->data[] = $row;
     }
 
-    print_heading(get_string('objectivescales', 'local_plan'));
-
-    $row = array();
-    $row[] = '';
-    $row[] = '<noscript><input type="submit" value="Update" /></noscript>';
-    $row[] = '';
-    $row[] = '';
-    $table->data[] = $row;
+    if($numvalues != 1) {
+        $row = array();
+        $row[] = '';
+        $row[] = '<input type="submit" value="Update" />';
+        $row[] = '';
+        $row[] = '';
+        $table->data[] = $row;
+    }
 
     print_table($table);
     echo "</form>\n";
-        ?>
-<script type="text/javascript">
-    $("#objscaleupdateform input:radio").change(
-        function(eventObject){
-            $("#objscaleupdateform").submit();
-        }
-    );
 
-    // On page load, remove last row in table (it's for non-js users only)
-    $(function() {
-        $('form#objscaleupdateform table.generaltable tr.lastrow').remove();
-    });
-</script>
-<?php
 } else {
     echo '<br /><div>'.get_string('noobjectivescalevalues','local_plan').'</div><br />';
 
@@ -315,8 +336,10 @@ if ($values) {
 echo '<div class="buttons">';
 
 // Print button for creating new objective scale value
-$options = array('objscaleid' => $objective->id);
-print_single_button($CFG->wwwroot.'/local/plan/objectivescales/editvalue.php', $options, get_string('addnewobjectivevalue', 'local_plan'), 'get');
+if(!$scale_used) {
+    $options = array('objscaleid' => $objective->id);
+    print_single_button($CFG->wwwroot.'/local/plan/objectivescales/editvalue.php', $options, get_string('addnewobjectivevalue', 'local_plan'), 'get');
+}
 
 echo '</div>';
